requirements.txt
gradio==3.40.0
huggingface_hub==0.16.4
transformers==4.34.0
diffusers==0.22.0
torch>=2.0.0
Pillow==9.5.0